// tdg-management-api-backend/src/projects/types/request.type.ts
export enum ProjectSortByOptions {
  DisplayOrderAsc = 'displayOrderAsc',
  EstimatedStartDateAsc = 'estimatedStartDateAsc',
  EstimatedStartDateDesc = 'estimatedStartDateDesc',
  StartDateAsc = 'startDateAsc',
  StartDateDesc = 'startDateDesc',
  CreatedAtAsc = 'createdAtAsc',
  CreatedAtDesc = 'createdAtDesc',
  EstimatedEndDateAsc = 'estimatedEndDateAsc',
  EstimatedEndDateDesc = 'estimatedEndDateDesc',
}